requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.5
django-filter==2.4.0
django-redis==4.12.1
django-simpleui==2021.4.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
install==1.3.4
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.0
PyJWT==1.7.1
PyMySQL==1.0.2
pysqlite3==0.4.5
python-ldap==3.3.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
